refactor : 등록 요구 검증함수 훅으로 분리
